feat: new components ans systems

+++ frontend/src/app/components/snackbar/snackbar.ts
@@ -1,6 +1,13 @@
 import { CommonModule } from "@angular/common";
-import { Component, OnInit } from "@angular/core";
+import { Component, Input, OnDestroy, OnInit } from "@angular/core";
 import { IconX } from "@assets/icons/x.icon";
+
+const IconVariants = {
+  'success': 'm424-408-86-86q-11-11-28-11t-28 11q-11 11-11 28t11 28l114 114q12 12 28 12t28-12l226-226q11-11 11-28t-11-28q-11-11-28-11t-28 11L424-408Zm56 328q-83 0-156-31.5T197-197q-54-54-85.5-127T80-480q0-83 31.5-156T197-763q54-54 127-85.5T480-880q83 0 156 31.5T763-763q54 54 85.5 127T880-480q0 83-31.5 156T763-197q-54 54-127 85.5T480-80Z',
+  'warning': 'M109-120q-11 0-20-5.5T75-140q-5-9-5.5-19.5T75-180l370-640q6-10 15.5-15t19.5-5q10 0 19.5 5t15.5 15l370 640q6 10 5.5 20.5T885-140q-5 9-14 14.5t-20 5.5H109Zm371-120q17 0 28.5-11.5T520-280q0-17-11.5-28.5T480-320q-17 0-28.5 11.5T440-280q0 17 11.5 28.5T480-240Zm0-120q17 0 28.5-11.5T520-400v-120q0-17-11.5-28.5T480-560q-17 0-28.5 11.5T440-520v120q0 17 11.5 28.5T480-360Z',
+  'error': 'M480-280q17 0 28.5-11.5T520-320q0-17-11.5-28.5T480-360q-17 0-28.5 11.5T440-320q0 17 11.5 28.5T480-280Zm0-160q17 0 28.5-11.5T520-480v-160q0-17-11.5-28.5T480-680q-17 0-28.5 11.5T440-640v160q0 17 11.5 28.5T480-440Zm0 360q-83 0-156-31.5T197-197q-54-54-85.5-127T80-480q0-83 31.5-156T197-763q54-54 127-85.5T480-880q83 0 156 31.5T763-763q54 54 85.5 127T880-480q0 83-31.5 156T763-197q-54 54-127 85.5T480-80Z',
+  'info': 'M480-280q17 0 28.5-11.5T520-320v-160q0-17-11.5-28.5T480-520q-17 0-28.5 11.5T440-480v160q0 17 11.5 28.5T480-280Zm0-320q17 0 28.5-11.5T520-640q0-17-11.5-28.5T480-680q-17 0-28.5 11.5T440-640q0 17 11.5 28.5T480-600Zm0 520q-83 0-156-31.5T197-197q-54-54-85.5-127T80-480q0-83 31.5-156T197-763q54-54 127-85.5T480-880q83 0 156 31.5T763-763q54 54 85.5 127T880-480q0 83-31.5 156T763-197q-54 54-127 85.5T480-80Z'
+} as const;
 
 @Component({
   selector: 'app-snackbar',
@@ -8,15 +15,20 @@ import { IconX } from "@assets/icons/x.icon";
   styleUrl: './snackbar.scss',
   imports: [CommonModule, IconX]
 })
-export class SnackbarComponent implements OnInit {
+export class Snackbar implements OnInit, OnDestroy {
 
   message: string = '';
 
   type: 'success' | 'warning' | 'error' | 'info' = 'info';
 
-  isVisible: boolean = false;
+  @Input()
+  duration?: number;
 
-  // We will assign this function from the service
+  isVisible = false;
+
+  isDestroyed = false
+
+  // Will assign this function from the service
   destroyRef: () => void = () => {};
 
   ngOnInit() {
@@ -24,13 +36,29 @@ export class SnackbarComponent implements OnInit {
     setTimeout(() => {
       this.isVisible = true
     }, 10);
+
+    if (this.duration && this.duration > 0) {
+      setTimeout(() => {
+        this.close();
+      }, this.duration);
+    }
+  }
+
+  ngOnDestroy(): void {
+    this.isDestroyed = true;
   }
 
   close() {
     this.isVisible = false;
     // Wait for animation to finish before destroying
     setTimeout(() => {
-      this.destroyRef();
+      if (!this.isDestroyed){
+        this.destroyRef();
+      }
     }, 300);
+  }
+
+  get iconPathD(): string {
+    return IconVariants[this.type as keyof typeof IconVariants]
   }
 }
